requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1

# Claude API & MCP
anthropic==0.39.0
mcp==1.1.0

# Async Support
httpx==0.27.2
aiohttp==3.10.10
# asyncio is built-in Python module, no need to install

# API Clients
requests==2.32.3
urllib3==2.2.3

# Data Processing
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
sqlite3-api==2.0.0
aiosqlite==0.20.0

# Natural Language Processing
spacy==3.8.2
nltk==3.9.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2
Pillow==11.0.0

# Caching
cachetools==5.5.0
redis==5.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-httpx==0.30.0

# Development Tools
black==24.10.0
flake8==7.1.1
mypy==1.13.0
pre-commit==4.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.3

# Logging & Monitoring
loguru==0.7.2
python-json-logger==2.0.7

# Rate Limiting
slowapi==0.1.9
limits==3.13.0

# Environment Management
python-decouple==3.8

# Optional: Production Server
gunicorn==23.0.0

# Optional: Background Tasks
celery==5.4.0
redis==5.2.0

# JSON Handling
orjson==3.10.11
ujson==5.10.0

# Markdown Support
markdown==3.7